requirements.txt
google-generativeai
Pillow
streamlit